Refactor Logout
